Add files via upload
This code reads in the CLASH and Molino catalogs and creates 3 sub-catalogs that include coordinates, ids, redshift, photometry, and photometry errors. Not all of the clumpy galaxies in CLASH's data set are included in Molino's dataset, so the CLASH sub-catalog will be longer. The third catalog keeps track of which galaxies are included in both datasets. This catalog includes a flag that indicates if the object is in both catalogs and includes the ids that both datasets use.