Add files via upload
This catalog contains the photometry for clumpy galaxies located in a383 and their individual clumps. The column labeled 'obj' indicates whether the photometry is for a single clump or the entire galaxy. An obj value of -1 is reserved for the entire galaxy, and any number greater than or equal to zero is a clump.